# openapi-dashboard.yaml
openapi: 3.0.3
info:
  title: ShopHub Dashboard API (Site-based)
  version: 1.0.0
  description: |
    API spec derived from the current dashboard UX flow:
    - Store Management Dashboard
    - Store Registration
    - Reviews page (/reviews)
    - Google Places latest reviews (latest 5)
    - AI reply drafting for recent reviews
servers:
  - url: https://api.shophub.kr
    description: production
  - url: http://localhost:8080
    description: local
...
paths:
...
  /v1/stores/{storeId}/reviews/summary:
    get:
      tags: [Reviews]
      summary: Get review summary for top section
      description: Returns values like "Google total 128, synced latest 5".
      parameters:
        - $ref: '#/components/parameters/StoreId'
        - in: query
          name: platform
          required: false
          schema:
            $ref: '#/components/schemas/PlatformType'
            default: GOOGLE
      responses:
        '200':
          description: Summary loaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewSummary'
  /v1/stores/{storeId}/reviews:
    get:
      tags: [Reviews]
      summary: Get reviews for reviews page
      description: Supports the recent-review list shown in /reviews.
      parameters:
        - $ref: '#/components/parameters/StoreId'
        - in: query
          name: platform
          schema:
            $ref: '#/components/schemas/PlatformType'
            default: GOOGLE
        - in: query
          name: syncedOnly
          schema:
            type: boolean
            default: true
        - in: query
          name: sort
          schema:
            type: string
            enum: [LATEST, RATING_DESC, RATING_ASC]
            default: LATEST
        - in: query
          name: page
          schema:
            type: integer
            default: 0
            minimum: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Reviews loaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'
...
components:
  parameters:
    StoreId:
      in: path
      name: storeId
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    PlatformType:
      type: string
      enum: [GOOGLE, INSTAGRAM, NAVER_PLACE, KAKAO_MAP, FACEBOOK, THREADS]
...
    ReviewSummary:
      type: object
      required: [platform, totalCount, syncedRecentCount]
      properties:
        platform:
          $ref: '#/components/schemas/PlatformType'
        totalCount:
          type: integer
          description: Total review count on source platform
        syncedRecentCount:
          type: integer
          description: Number of reviews currently synced into ShopHub
        averageRating:
          type: number
          format: double
        externalViewUrl:
          type: string
          format: uri
    Review:
      type: object
      required: [id, storeId, platform, authorName, rating, content, reviewedAt]
      properties:
        id:
          type: string
          format: uuid
        storeId:
          type: string
          format: uuid
        platform:
          $ref: '#/components/schemas/PlatformType'
        externalReviewId:
          type: string
        authorName:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        content:
          type: string
        reviewedAt:
          type: string
          format: date-time
        replied:
          type: boolean
        reply:
          $ref: '#/components/schemas/ReviewReply'
    ReviewListResponse:
      type: object
      required: [items, page, size, totalElements]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
...
    ReviewReply:
      type: object
      required: [id, content, createdAt]
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time